+++ app/client/event_listeners/event_listeners.js
@@ -11,10 +11,7 @@ import { listener_exclude_button } from "./listener_exclude_button.js";
 import { listener_toggle_layers } from "./listener_toggle_layers.js";
 import { listener_toggle_blacklist } from "./listener_toggle_blacklist.js";
 import { listener_show_crimes } from "./listener_show_crimes.js";
-import { listener_show_supermarkets } from "./listener_show_supermarkets.js";
-import { listener_show_schools } from "./listener_show_schools.js";
-import { listener_show_convenience } from "./listener_show_convenience.js";
-import { listener_show_coffee } from "./listener_show_coffee.js";
+import { listener_show_openpass } from "./listener_show_openpass.js";
 import { listener_blacklist_item } from "./listener_blacklist_item.js";
 import { listener_blacklist_item_remove } from "./listener_blacklist_item_remove.js";
 import { listener_inspector_text } from "./listener_inspector_text.js";
@@ -49,10 +46,6 @@ export function load_event_listeners(){
     listener_toggle_layers()
     listener_toggle_blacklist()
     listener_show_crimes()
-    listener_show_supermarkets()
-    listener_show_schools()
-    listener_show_convenience()
-    listener_show_coffee()
     listener_blacklist_item()
     listener_blacklist_item_remove()
     listener_inspector_text()
@@ -78,4 +71,115 @@ export function load_event_listeners(){
     filter_invert()
     filter_dark()
     filter_sepia()
+
+    // OpenPass layers
+    // GYMS
+    listener_show_openpass({
+        'layerID':      'local-gyms',
+        'buttonID':     'load-local-gyms',
+        'openpassKey':  'leisure',
+        'openpassValue':'fitness_centre',
+        'fill':         'rgba(30,30,30, 1.0)',
+        'stroke':       'rgba(120, 0, 0, 1.0)',
+    })
+
+    // Convenience stores
+    listener_show_openpass({
+        'layerID':      'local-convenience',
+        'buttonID':     'load-local-convenience',
+        'openpassKey':  'shop',
+        'openpassValue':'convenience',
+        'fill':         'rgba(255, 255, 255, 1.0)',
+        'stroke':       'rgba(0, 0, 0, 1.0)',
+    })
+
+    // Coffee
+    listener_show_openpass({
+        'layerID':      'local-coffee',
+        'buttonID':     'load-local-coffee',
+        'openpassKey':  'amenity',
+        'openpassValue':'cafe',
+        'fill':         'rgba(120, 53, 15, 1.0)',
+        'stroke':       'rgba(245, 158, 11, 1.0)',
+    })
+
+    // Schools
+    listener_show_openpass({
+        'layerID':      'local-school',
+        'buttonID':     'load-local-schools',
+        'openpassKey':  'amenity',
+        'openpassValue':'school',
+        'fill':         'rgba(252, 211, 77, 1.0)',
+        'stroke':       'rgba(245, 245, 245, 1.0)',
+    })
+
+    // Supermarkets
+    listener_show_openpass({
+        'layerID':      'local-supermarkets',
+        'buttonID':     'load-local-supermarkets',
+        'openpassKey':  'shop',
+        'openpassValue':'supermarket',
+        'fill':         'rgba(255, 255, 255, 1.0)',
+        'stroke':       'rgba(0, 0, 0, 1.0)',
+    })
+
+    // Post-Office
+    listener_show_openpass({
+        'layerID':      'local-post-office',
+        'buttonID':     'load-local-post-office',
+        'openpassKey':  'amenity',
+        'openpassValue':'post_office',
+        'fill':         'rgba(255, 0, 0, 1.0)',
+        'stroke':       'rgba(0, 0, 0, 1.0)',
+    })
+
+    // Train/Tube Station
+    listener_show_openpass({
+        'layerID':      'local-station',
+        'buttonID':     'load-local-station',
+        'openpassKey':  'public_transport',
+        'openpassValue':'station',
+        'fill':         'rgba(0, 0, 240, 1.0)',
+        'stroke':       'rgba(255, 255, 255, 1.0)',
+    })
+
+    // Bus Stop
+    listener_show_openpass({
+        'layerID':      'local-bus-stop',
+        'buttonID':     'load-local-bus-stop',
+        'openpassKey':  'public_transport',
+        'openpassValue':'platform',
+        'fill':         'rgba(0, 240, 0, 1.0)',
+        'stroke':       'rgba(255, 255, 255, 1.0)',
+    })
+
+    // Airport
+    listener_show_openpass({
+        'layerID':      'local-airport',
+        'buttonID':     'load-local-airport',
+        'openpassKey':  'aeroway',
+        'openpassValue':'aerodrome',
+        'fill':         'rgba(120, 120, 0, 1.0)',
+        'stroke':       'rgba(5, 150, 255, 1.0)',
+    })
+
+    // Doctor
+    listener_show_openpass({
+        'layerID':      'local-doctor',
+        'buttonID':     'load-local-doctor',
+        'openpassKey':  'amenity',
+        'openpassValue':'doctors',
+        'fill':         'rgba(255, 255, 255, 1.0)',
+        'stroke':       'rgba(230, 0, 0, 1.0)',
+    })
+
+    // Hospital
+    listener_show_openpass({
+        'layerID':      'local-hospital',
+        'buttonID':     'load-local-hospital',
+        'openpassKey':  'amenity',
+        'openpassValue':'hospital',
+        'fill':         'rgba(255, 200, 200, 1.0)',
+        'stroke':       'rgba(230, 0, 0, 1.0)',
+    })
 }